requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
django-environ==0.12.0
django-apscheduler==0.7.0
APScheduler==3.11.0
sentry-sdk~=2.25.1